+++ cruddy/model.py
@@ -1,16 +1,18 @@
 """ database dependencies to support Users db examples """
-from sqlalchemy.exc import IntegrityError
 from __init__ import db
+from sqlalchemy.exc import IntegrityError
+from werkzeug.security import generate_password_hash, check_password_hash
+from flask_login import UserMixin
+
 
 # Tutorial: https://www.sqlalchemy.org/library.html#tutorials, try to get into Python shell and follow along
-
 
 # Define the Users table within the model
 # -- Object Relational Mapping (ORM) is the key concept of SQLAlchemy
 # -- a.) db.Model is like an inner layer of the onion in ORM
 # -- b.) Users represents data we want to store, something that is built on db.Model
 # -- c.) SQLAlchemy ORM is layer on top of SQLAlchemy Core, then SQLAlchemy engine, SQL
-class Users(db.Model):
+class Users(UserMixin, db.Model):
     # define the Users schema
     userID = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(255), unique=False, nullable=False)
@@ -22,7 +24,7 @@ class Users(db.Model):
     def __init__(self, name, email, password, phone):
         self.name = name
         self.email = email
-        self.password = password
+        self.set_password(password)
         self.phone = phone
 
     # CRUD create/add a new record to the table
@@ -56,7 +58,7 @@ class Users(db.Model):
         if len(name) > 0:
             self.name = name
         if len(password) > 0:
-            self.password = password
+            self.set_password(password)
         if len(phone) > 0:
             self.phone = phone
         db.session.commit()
@@ -68,6 +70,21 @@ class Users(db.Model):
         db.session.delete(self)
         db.session.commit()
         return None
+
+    # set password method is used to create encrypted password
+    def set_password(self, password):
+        """Create hashed password."""
+        self.password = generate_password_hash(password, method='sha256')
+
+    # check password to check versus encrypted password
+    def is_password_match(self, password):
+        """Check hashed password."""
+        result = check_password_hash(self.password, password)
+        return result
+
+    # required for login_user, overrides id (login_user default) to implemented userID
+    def get_id(self):
+        return self.userID
 
 
 """Database Creation and Testing section"""
