+++ main.py
@@ -2,11 +2,14 @@ from flask import Flask, render_template, request
 from __init__ import app
 import requests
 from cruddy.app_crud import app_crud
+from cruddy.app_crud_api import app_crud_api
 
 
 from pathlib import Path
 
 app.register_blueprint(app_crud)
+app.register_blueprint(app_crud_api)
+
 
 
 # connects default URL to render index.html
